started implementing game close/suspend; fix errors; general cleanup

+++ source/Services/Hid.cpp
@@ -27,14 +27,14 @@ bool Hid::IsTouched(){
 bool Hid::IsTouched(SDL_Rect pos){
     touchPosition touchPos;
     hidTouchRead(&touchPos, 0);
-    return (touchPos.px >= pos.x && touchPos.px <= (pos.x + pos.w) && touchPos.py >= pos.y && touchPos.py <= (pos.y + pos.h));
+    return ((int)touchPos.px >= pos.x && (int)touchPos.px <= (pos.x + pos.w) && (int)touchPos.py >= pos.y && (int)touchPos.py <= (pos.y + pos.h));
 }
 
 bool Hid::IsTapped(SDL_Rect pos){
     touchPosition touchPos;
     hidTouchRead(&touchPos, 0);
 	int i = svcGetSystemTick();
-    while(touchPos.px >= pos.x && touchPos.px <= (pos.x + pos.w) && touchPos.py >= pos.y && touchPos.py <= (pos.y + pos.h)) {
+    while((int)touchPos.px >= pos.x && (int)touchPos.px <= (pos.x + pos.w) && (int)touchPos.py >= pos.y && (int)touchPos.py <= (pos.y + pos.h)) {
 		if(svcGetSystemTick() - i >= SHORTPRESS_MS) return false;
 		hidTouchRead(&touchPos, 0);
 	}
